memorymanager: throw a mutex around this structure
There *ought* to only be one user of this structure at a time, however
humans are sloppy.  Throw a cheesy mutex around it in order to ensure we
don't reuse this structure.

Note that we use non-atomic operations.  This could be a problem if we
ran multicore or accessed the memory manager from the main thread,
however we only access it from interupt contexts and we're currently
single-core, so it's not an issue.

Signed-off-by: Sean Cross <sean@xobs.io>