systemservices: place struct behind handle
This patch places the SystemServices struct behind a
SystemServicesHandle, which performs a rudimentary form of reference
counting.  This is used to panic when there is more than one user of the
SystemServices struct at once.

As a result of this change, many conflicts were discovered, so some
major rework needed to be done.  We now have a unified exit from the
interrupt handler, and we have limited the number of process switching
routines to two.

Signed-off-by: Sean Cross <sean@xobs.io>